src/requirements.txt
# unikube commons
git+git://github.com/unikubehq/commons.git@1.3.1

# regular pypi
bpython
Django>=2.0,<3.0
django-extensions
django-hurricane~=0.8.8
environs~=8.0.0
gitpython~=3.1.9
graphene-django==2.9.1
graphene-federation~=0.1.0
psycopg2>=2.7,<3.0

# tests
factory_boy